+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Postman / Newman integration tests to ensure giuseppe works with the given express version.
+
 ### Fixed
 - When using `~` on a controller and a route, it uses the correct root for the route now.